Changing an approach how click calculation is performed

+++ Reactive Programming/Assets/Scripts/Types/StructureInteraction.cs
@@ -2,6 +2,6 @@
 namespace Types.Enums {
      public struct StructureInteraction {
           public GovernmentInteractionType GovernmentInteraction;
-          public int InteractionResult;
+          public long InteractionResult;
      }
 }
